se crea funcion y solicitud  eliminar usuarios
